requirements.txt
playwright
python-dotenv
pandas
fastapi
uvicorn
sqlalchemy
psycopg[binary]